src/Architecture.Template.Application/TodoItem/Commands/CreateTodoItem/CreateTodoItemCommandHandler.cs
using Domain.Events;
using Domain.Interfaces.Repository;

namespace Application.TodoItem.Commands.CreateTodoItem;
public class CreateTodoItemCommandHandler : IRequestHandler<CreateTodoItemCommand, Guid>
{
    private readonly ITodoItemRepository _todoItemRepository;

    public CreateTodoItemCommandHandler(ITodoItemRepository todoItemRepository) =>
        _todoItemRepository = todoItemRepository ?? throw new ArgumentNullException(nameof(todoItemRepository));

    public async Task<Guid> Handle(CreateTodoItemCommand request, CancellationToken cancellationToken)
    {
        var entity = new TodoItemEntity
        {
            ListId = request.ListId,
            Title = request.Title,
            Done = false
        };

        entity.AddDomainEvent(new TodoItemCreatedEvent(entity));

        await _todoItemRepository.InsertAsync(entity, cancellationToken);

        return entity.Id;
    }
}
